docs/Reference/e-gmd (analysis)/requirements.txt
ijson>=3.1.4
tqdm>=4.0.0
mido>=1.2.10
numpy>=1.21.0
pyyaml>=6.0
